+++ src/app/admin/admin-woodburnings/admin-woodburnings.component.scss
@@ -1,4 +1,4 @@
-@import '../../common/colors';
+@import '../../../common/colors';
 
 .woodburning-table {
   margin-top: 10px;
