+++ src/app/nav/nav.component.html
@@ -24,7 +24,8 @@
                   <mat-icon>lock</mat-icon> Admin<span class="caret"></span>
                 </a>
                 <div *dropdownMenu class="dropdown-menu dropdown-menu-right dropdown dropdown-primary" role="menu">
-                  <a class="dropdown-item waves-light" mdbWavesEffect routerLink="admin-home" [routerLinkActive]="['active']">
+                  <a class="dropdown-item waves-light" mdbWavesEffect routerLink="admin-woodburnings"
+                     [routerLinkActive]="['active']">
                       <mat-icon>work</mat-icon> Manage Woodburnings
                   </a>
                   <a class="dropdown-item waves-light" mdbWavesEffect routerLink="admin-map-locations" [routerLinkActive]="['active']">
